refactor reduce and make it compatible with >65k tokens
