+++ aiter/ops/triton/_triton_kernels/moe/reduce.py
@@ -1,0 +1,76 @@
+import torch
+import triton
+import triton.language as tl
+from aiter.ops.triton._triton_kernels.moe.activations import _swiglu
+
+
+@triton.jit
+def _reduce_grouped(
+    X,
+    stride_xb: tl.uint64,
+    stride_xm: tl.uint64,
+    stride_xn,  #
+    Out,
+    stride_om: tl.uint64,
+    stride_on,  # output tensor
+    InIndx,
+    B,
+    N,
+    num_blocks,
+    # fused activation function
+    APPLY_SWIGLU: tl.constexpr,
+    alpha,
+    limit,
+    ACTIVATION_REDUCTION_N: tl.constexpr,
+    K: tl.constexpr,
+    BLOCK_N: tl.constexpr,
+    EVEN_N: tl.constexpr,
+    ADD_RESIDUAL: tl.constexpr,
+):
+    pid = tl.program_id(0)
+    pid_t = pid // num_blocks
+    pid_n = pid % num_blocks
+
+    BLOCK_N_OUT: tl.constexpr = BLOCK_N // ACTIVATION_REDUCTION_N
+    start = pid_t * K
+    # load indices into a tuple
+    if InIndx is None:
+        indxs = (pid_t,)
+    else:
+        indxs = ()
+        for i in tl.static_range(0, K):
+            indxs = indxs + (tl.load(InIndx + start + i),)
+    XPtrs = X + (pid_n * BLOCK_N + tl.arange(0, BLOCK_N)) * stride_xn
+    OutPtrs = Out + (pid_n * BLOCK_N_OUT + tl.arange(0, BLOCK_N_OUT)) * stride_on
+
+    acc = tl.zeros([BLOCK_N_OUT], dtype=tl.float32)
+    x_n_mask = pid_n * BLOCK_N + tl.arange(0, BLOCK_N) < N
+    # accumulate contributions for this tile
+    for i in tl.static_range(0, K):
+        curr = tl.zeros([BLOCK_N], dtype=tl.float32)
+        # iterate over split_k partial values
+        for b in tl.range(0, B):
+            x_row_ptr = XPtrs + indxs[i] * stride_xm + b * stride_xb
+            if EVEN_N:
+                vals = tl.load(x_row_ptr)
+            else:
+                vals = tl.load(x_row_ptr, mask=x_n_mask, other=0.0)
+            vals = vals.to(tl.float32)
+            curr += vals
+
+        # apply nonlinearity to split-k output
+        if APPLY_SWIGLU:
+            curr = _swiglu(curr[None, :], alpha, limit, ADD_RESIDUAL=ADD_RESIDUAL)
+        curr = tl.reshape(curr, [curr.shape[-1]])
+        # update final accumulator
+        acc += curr
+    # Compute per-32-col MXFP scales for this tile if requested
+    Nrem = N // ACTIVATION_REDUCTION_N
+
+    # write-back for this tile
+    out_ptr = OutPtrs + pid_t * stride_om
+    if EVEN_N:
+        tl.store(out_ptr, acc)
+    else:
+        out_n_mask = pid_n * BLOCK_N_OUT + tl.arange(0, BLOCK_N_OUT) < Nrem
+        tl.store(out_ptr, acc, mask=out_n_mask)
