+++ aiter/ops/triton/moe/moe_op_gemm_int8_smoothquant.py
@@ -8,8 +8,9 @@ from aiter.ops.triton.moe.moe_routing.routing import RoutingData
 from aiter.ops.triton.utils.device_info import get_num_sms
 from aiter.ops.triton._triton_kernels.moe.moe_op_gemm_int8_smoothquant import (
     _moe_gemm_int8_smoothquant,
-    _reduce_grouped,
 )
+from aiter.ops.triton.moe.reduce import reduce_grouped
+from aiter.ops.triton.utils._triton.arch_info import get_arch
 
 # -----------------------------------------------------------------------------
 #                    Matrix Multiplication + Outer Gather/Scatter
@@ -114,7 +115,7 @@ def get_kernel_config(m, n, k, routing_data):
         block_k = 256
         num_warps = 4
         num_stages = 2
-        kpack = 2
+        kpack = 2 if get_arch() == "gfx942" else 1
 
         grid_m = routing_data.n_blocks(m, block_m)
         grid_n = triton.cdiv(n, block_n)
@@ -145,83 +146,6 @@ def get_kernel_config(m, n, k, routing_data):
         "kpack": kpack,
     }
     return ret
-
-
-def reduce_grouped(
-    x: torch.Tensor,
-    indx: torch.Tensor,
-    out: torch.Tensor,
-    alpha=1.0,
-    limit=1.0,
-    reduction_n=1,
-    apply_activation: bool = False,
-    out_dtype: torch.dtype = None,
-    add_residual: bool = False,
-):
-    """
-    In-place grouped row reduction.
-
-    Arguments
-    - x: Tensor[AnyFloat] of shape [(num_groups * K), N]
-    - indx: Tensor[Int] of shape [num_groups, K]
-
-    Description
-    For each group g in [0, num_groups), this routine sums the K rows of `x`
-    specified by `indx[g, :]` and overwrites the row corresponding to the first
-    valid (non-negative) index with the per-group sum. Accumulation is performed
-    in float32 for numerical stability, and the result is written back in the
-    dtype of `x`.
-
-    Behavior and edge cases
-    - Invalid (-1) entries are skipped during accumulation and do not generate
-      memory traffic. If a group has no valid entries, nothing is written for
-      that group.
-    - Reduction is performed tile-by-tile along the N dimension within a single
-      kernel launch (persistent along N) to minimize launch overhead.
-
-    Performance notes
-    - Memory traffic per group is approximately (valid_rows_read + 1) * N * sizeof(x),
-      plus index reads. With no invalid entries, this becomes (K + 1) reads/writes
-      of length N per group.
-
-    Returns
-    - The input tensor `x` (modified in place).
-    """
-    if indx is None and x.shape[0] == 1:
-        return x.squeeze(0)
-
-    if indx is not None:
-        num_groups = indx.shape[0]
-    else:
-        num_groups = x.shape[-2]
-    K = 1 if indx is None else indx.shape[1]
-    out_dtype = x.dtype if out_dtype is None else out_dtype
-    assert x.shape[-1] % reduction_n == 0
-    BLOCK_N = 512
-    num_blocks = triton.cdiv(x.shape[-1], BLOCK_N)
-
-    _reduce_grouped[(num_blocks, num_groups)](
-        x,
-        x.stride(0),
-        x.stride(1),
-        x.stride(2),
-        out,
-        out.stride(0),
-        out.stride(1),
-        indx,
-        x.shape[0],
-        x.shape[-1],
-        alpha,
-        limit,
-        reduction_n,
-        BLOCK_N=BLOCK_N,
-        EVEN_N=(x.shape[-1] % BLOCK_N == 0),
-        K=K,
-        num_warps=2,
-        ADD_RESIDUAL=add_residual,
-        APPLY_ACTIVATION=apply_activation,
-    )
-    return out
 
 
 # -----------------------------------------------------------------------------
@@ -372,11 +296,10 @@ def moe_gemm_int8_smoothquant(
         y,
         group_indx,
         y_final,
+        apply_activation and (config["split_k"] > 1),  # apply activation if split_k > 1
         alpha,
         limit,
         reduction_n_reduction,
-        apply_activation=(alpha != 0)
-        and (config["split_k"] > 1),  # apply activation if split_k > 1
         out_dtype=out_dtype,
         add_residual=add_residual,
     )
